move decode interface posision
