+++ support/iam/src/console_tenant/api/iam_ct_cert_manage_api.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use bios_sdk_invoke::clients::spi_log_client::{LogDynamicContentReq, SpiLogClient};
 use tardis::web::context_extractor::TardisContextExtractor;
 use tardis::web::poem_openapi;
@@ -11,7 +9,7 @@ use bios_basic::rbum::dto::rbum_cert_dto::{RbumCertDetailResp, RbumCertSummaryWi
 use bios_basic::rbum::dto::rbum_filer_dto::RbumCertFilterReq;
 use bios_basic::rbum::dto::rbum_rel_dto::RbumRelBoneResp;
 
-use crate::basic::dto::iam_cert_dto::{IamCertDecodeRequest, IamCertManageAddReq, IamCertManageModifyReq};
+use crate::basic::dto::iam_cert_dto::{IamCertManageAddReq, IamCertManageModifyReq};
 use crate::basic::serv::iam_cert_serv::IamCertServ;
 use crate::iam_constants;
 use crate::iam_enumeration::IamCertExtKind;
@@ -219,13 +217,5 @@ impl IamCtCertManageApi {
         TardisResp::ok(rbum_certs)
     }
 
-    /// decode cert
-    #[oai(path = "/decode", method = "post")]
-    async fn decode_certs(&self, body: Json<IamCertDecodeRequest>, ctx: TardisContextExtractor, request: &Request) -> TardisApiResult<HashMap<String, String>> {
-        let funs = iam_constants::get_tardis_inst();
-        let ctx = IamCertServ::use_sys_or_tenant_ctx_unsafe(ctx.0)?;
-        add_remote_ip(request, &ctx).await?;
-        let doceded = IamCertServ::batch_decode_cert(body.0.codes, &funs, &ctx).await?;
-        TardisResp::ok(doceded)
-    }
+
 }
